refactor: Finalize move to modules with .mjs and correct imports
