+++ scripts/modules/slider.mjs
